requirements.txt
azure-functions
azure-cosmos
azure-storage
azure-cli
azure-apim